# docs/api.yaml
openapi: 3.0.0
info:
  title: IBUC System API
  description: API REST para o Sistema de Gestão de Curso de Teologia Infanto-Juvenil
  version: 1.0.0
  contact:
    name: IBUC - Instituto Bíblico Único Caminho
    email: contato@ibuc.com.br
servers:
  - url: https://seu-projeto.supabase.co/rest/v1
    description: Supabase REST API
security:
  - bearerAuth: []
paths:
...
  /matriculas:
    get:
      summary: Lista matrículas
      description: Retorna lista de matrículas respeitando RLS
      tags:
        - Matrículas
      parameters:
        - name: polo_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, em_analise, ativa, recusada, cancelada]
      responses:
        '200':
          description: Lista de matrículas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Matricula'
    post:
      summary: Cria uma nova matrícula
      description: Cria matrícula (pré-matrícula online ou presencial)
      tags:
        - Matrículas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatriculaInput'
      responses:
        '201':
          description: Matrícula criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Matricula'
...
  /matriculas/protocolo/{protocolo}:
    get:
      summary: Busca matrícula por protocolo
      description: Busca matrícula usando o protocolo (público)
      tags:
        - Matrículas
      parameters:
        - name: protocolo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados da matrícula
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Matricula'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Matricula:
      type: object
      properties:
        id:
          type: string
          format: uuid
        aluno_id:
          type: string
          format: uuid
        polo_id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [online, presencial]
        status:
          type: string
          enum: [pendente, em_analise, ativa, recusada, cancelada]
        protocolo:
          type: string
        created_at:
          type: string
          format: date-time
    MatriculaInput:
      type: object
      required:
        - aluno_id
        - polo_id
        - tipo
      properties:
        aluno_id:
          type: string
          format: uuid
        polo_id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [online, presencial]
...
  responses:
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Not found"